refactor: move component and routes to dashboard
