+++ app/javascript/dashboard/components/form/Input.vue
