+++ app/javascript/dashboard/routes/dashboard/start/components/ModalLayout.vue
