+++ app/javascript/v3/views/routes.js
@@ -5,9 +5,6 @@ const Signup = () => import('./auth/signup/Index.vue');
 const ResetPassword = () => import('./auth/reset/password/Index.vue');
 const Confirmation = () => import('./auth/confirmation/Index.vue');
 const PasswordEdit = () => import('./auth/password/Edit.vue');
-const SetupProfile = () => import('./auth/start/SetupProfile.vue');
-const SetupCompany = () => import('./auth/start/SetupCompany.vue');
-const FounderNote = () => import('./auth/start/FounderNote.vue');
 
 export default [
   {
@@ -55,20 +52,5 @@ export default [
     path: frontendURL('auth/reset/password'),
     name: 'auth_reset_password',
     component: ResetPassword,
-  },
-  {
-    path: frontendURL('start/setup-profile'),
-    name: 'onboarding_setup_profile',
-    component: SetupProfile,
-  },
-  {
-    path: frontendURL('start/setup-company'),
-    name: 'onboarding_setup_company',
-    component: SetupCompany,
-  },
-  {
-    path: frontendURL('start/founders-note'),
-    name: 'onboarding_founders_note',
-    component: FounderNote,
   },
 ];
